refactor(desktop): tighten Omarchy theme integration
