+++ apps/web/src/routes/__root.tsx
@@ -17,11 +17,11 @@ import { CommandPalette } from "../components/CommandPalette";
 import { ConfirmDialogHost } from "../components/ConfirmDialogHost";
 import { ConnectOnboardingDialog } from "../components/cloud/ConnectOnboardingDialog";
 import { RelayClientInstallDialog } from "../components/cloud/RelayClientInstallDialog";
+import { OmarchyLinkedThemeSync } from "../components/desktop/OmarchyLinkedThemeSync";
 import { SshPasswordPromptDialog } from "../components/desktop/SshPasswordPromptDialog";
 import { ProviderUpdateLaunchNotification } from "../components/ProviderUpdateLaunchNotification";
 import { SlowRpcRequestToastCoordinator } from "../components/SlowRpcRequestToastCoordinator";
 import { ThemeEditorHost } from "../components/settings/ThemeEditorHost";
-import { OmarchyLinkedThemeSync } from "../components/settings/OmarchyLinkedThemeSync";
 import { Button } from "../components/ui/button";
 import {
   AnchoredToastProvider,
